% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp3c.r
\name{gtrack.create_redb_tracks}
\alias{gtrack.create_redb_tracks}
\title{Create a new redb for UMI-4C}
\usage{
gtrack.create_redb_tracks(re_seq, redb_params_fn, verbose = TRUE)
}
\arguments{
\item{re_seq}{sequence recognition (e.g. 'GATC' for DpnII)}

\item{redb_params_fn}{File with the redb parameters ('redb.conf')}
}
\description{
Generate a new database for a restriction enzyme in the redb.conf file.
}
