% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp3c.r
\name{p4cBuildRequirements}
\alias{p4cBuildRequirements}
\title{Build the required genomic datasets.}
\usage{
p4cBuildRequirements(output_dir, genome = "hg19", reseq = "GATC",
  dataset_keys = "http://www.wisdom.weizmann.ac.il/~omersch/dataset_keys.txt")
}
\arguments{
\item{output_dir}{Output directory for the dataset.}

\item{genome}{For which genome build should the data be download. Current options available are:
'hg19', 'mm9', 'mm10'.}

\item{reseq}{Restriction enzye database. Currently only DpnII (\strong{GATC}) is available.}
}
\description{
Downloads the required datasets for the package: trackdb and redb. These datasets
include a genomic database in a binary format and restriction sites tables in text format.
}
