% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp3c.r
\name{p4cCreate4CseqTrack}
\alias{p4cCreate4CseqTrack}
\title{Import fastq files to a genomic track.}
\usage{
p4cCreate4CseqTrack(sample_ids, track_desc = "4C track", verbose = TRUE,
  skip.if.exist = FALSE, overwrite.if.exist = TRUE, groot = NULL)
}
\arguments{
\item{sample_ids}{Numerical samples IDs from the baits.txt table.}

\item{track_desc}{Description attribute. Optional.}

\item{verbose}{Logical. Controls verbosity.}

\item{groot}{Optional. Controls the trackdb to which tracks are imported.}

\item{skip.if.exists}{Should it skip if it sees an exisiting tack with the same name?}

\item{overwrite.if.exists}{Should a new track overwrite an older one with the same name?}
}
\description{
Generate a genomic track from fastqs listed in the samples.conf file.
}

