% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp3c.r
\name{p4cDumpConfFiles}
\alias{p4cDumpConfFiles}
\title{Dump config files templates.}
\usage{
p4cDumpConfFiles(conf_dir)
}
\arguments{
\item{conf_dir}{Directory to dump to files to.}
}
\description{
Dump templates of config files required by the package.
}
