% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4c_package.R
\name{p4cIntervalsMean}
\alias{p4cIntervalsMean}
\title{Compare contacts intensities between two profiles}
\usage{
p4cIntervalsMean(p4c_obj, ref_p4c_obj, start, end, min_win_cov = 30,
  rawdata = FALSE, max_scope = 1000000)
}
\arguments{
\item{p4c_obj}{A \code{p4cProfile} object.}

\item{ref_p4c_obj}{The \code{p4cProfile} object one wishes to compare to.}

\item{start, end}{Start and end coordinates of the genomic interval.}

\item{min_win_cov}{Controls the smoothing as described in \code{plot.p4cProfile}.}

\item{rowdata}{Return raw molecule counts instead of smoothed data.}
}
\description{
Compare contact intensities at specific genomic intervals.
}
\details{
When a genomic region is suspected to be defferentialy contacting between two profiles, 
it is possible to compare these two profiles in the same intervals by deriving the normalized contact values
in that region. This function facilitates this task. It returns a data.frame with 
the either the row molecule counts if \code{rowdata = TRUE} or normalized other wise.
The smoothing is controlled by \code{min_win_cov}.

The p-value is calculated for the entire interval with a chi-square test. However,
it assumes the interval is <1Mb from the bait. If the interval is >1Mb from the bait
 \code{max_scope} should be increased accordingly.
}

