% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp3c.r
\name{p4cLoadConfFiles}
\alias{p4cLoadConfFiles}
\title{Load all the config files required for the package.}
\usage{
p4cLoadConfFiles(
  conf_dir,
  paths_conf = file.path(conf_dir, "paths.conf"),
  pipeline_conf = file.path(conf_dir, "pipeline.conf"),
  p4c_conf = file.path(conf_dir, "p4c.conf")
)
}
\arguments{
\item{conf_dir}{A directory with all the defined conf files. Should have the same
files as those that were exported with \code{p4cDumpConfFiles}.}

\item{paths_conf}{Parameter file like \emph{paths.conf}}

\item{pipeline_conf}{Parameter file like \emph{pipeline.conf}}

\item{p4c_conf}{Parameter file like \emph{p4c.conf}}
}
\description{
Loads the config files which were filled by the user. Accepts either directory path
from \code{conf_dir}, or individual files supplied by the other parameters.
}
