% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4c_package.R
\name{p4cNewProfile}
\alias{p4cNewProfile}
\title{Create a new p4cProfile object.}
\usage{
p4cNewProfile(
  track_nm,
  bait_chrom = gtrack.attr.get(track_nm[1], "Bait_chr"),
  bait_start = as.numeric(gtrack.attr.get(track_nm[1], "Bait_coord")),
  scope_5 = 250000,
  scope_3 = 250000,
  stat_type
)
}
\arguments{
\item{track_nm}{2D genomic track name. Can also be a character vector of track
names that will be pooled together by summation.}

\item{bait_chrom}{Character. Bait chromosome.}

\item{bait_start}{Integer. Start coordinate of the bait.}

\item{scope_5, scope_3}{Define the scope of the profile. How many bases upstream
and downstream of the bait should be used for generating the profile. 
Default is set to 250kb upstream and downstream.}

\item{stat_type}{Define the method for calculating the profile contact intensities. 
Either "linear" or "log", for arithmetic or geometric means, respectively. Default 
is "linear".}
}
\value{
Returns a p4cProfile object.
}
\description{
\code{p4cNewProfile} constructs a p4cProfile object from a genomic 2D track.
}
\details{
This function is the primary object constructor. It processes a 2D genomic track 
with (UMI)-4C raw contact data and returns a p4cProfile object.
}
