% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp3c.r
\name{p4cCreate4CseqTrack}
\alias{p4cCreate4CseqTrack}
\title{Import fastq files to a genomic track.}
\usage{
p4cCreate4CseqTrack(sample_ids = NULL, track_desc = "4C track",
  verbose = TRUE, overwrite.if.exist = FALSE, groot = NULL)
}
\arguments{
\item{sample_ids}{Numerical samples IDs from the baits.txt table.}

\item{track_desc}{Description attribute. Optional.}

\item{verbose}{Logical. Controls verbosity.}

\item{groot}{Optional. Controls the trackdb to which tracks are imported.}

\item{overwrite.if.exists}{Force overwriting on older tracks.}
}
\description{
Generate a genomic track from fastqs listed in the samples.conf file.
}
\examples{
\donttest{
# Load config files
p4cLoadConfFiles("conf/")

# Import 3 tracks
p4cCreate4CseqTrack(1:3)
}

}
