% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4c_package.R
\name{p4cExportBedGraph}
\alias{p4cExportBedGraph}
\title{Export bedGraph}
\usage{
p4cExportBedGraph(p4c_obj, filename, min_win_cov = 50, color = "black")
}
\arguments{
\item{p4c_obj}{\code{p4cProfile} object}

\item{filename}{Filename of the bedGraph.}

\item{min_win_cov}{Smoothing parameter.}

\item{color}{Color of the bedGraph plot.}
}
\description{
\code{p4cExportBedGraph} exports the smoothed trend to a file which can be imported as a 'custom track'
to the UCSC genome browser.
}
